Delete parameter: getfrequentrenterpoints
When extracting the method, we received the total number of frequent renter
points, modified it, and returned it. It makes more sense to calculate
the individual amount of points each rental is worth, so we rework this
method to return this value instead, and add it to the total in the
original loop instead.